Table changes, hide .jsp pages, and redirecting.
At this stage, I changed a little data base, and exactly I deleted
tables with obligatory holidays, which will be needed to consider it in
generated final schedule but we will define it in code. Also I added
first name, and last name of employees to the table users. Except that,
I also hide .jsp pages that users will not be able to refer to them
directly, and also I made all redirecting related to those
pages/servlets.